code improvements to make it easier to read
